+++ src/db/schema/mysql-schema.sql
@@ -569,7 +569,7 @@ CREATE TABLE t_file (
   
 --
 -- user provided size of the file (bytes)
-  user_filesize  BIGINT,  
+  user_filesize  DOUBLE,  
   
 --
 -- File metadata
